Add normalizer instead of abusing LiteLLM adapter
There were 2 problems with how our classes were structured (there are
more but these two I attempted to solve with this PR):
1) The adapter was supposed to be used in the litellm based providers
   and was supposed to do the translation using litellm's adapters. But
   we stuffed way too much logic in them and started leaking the logic
   to other providers
2) Despite litellm using the openaAI format for input and output, other
   providers (llama.cpp and soon to be hosted vllm) don't. We need a way
   to canonicalize them to openAI format.

This PR adds a new module called normalizer that takes some of the work
from the adapter and is only responsible for changing the requests and
replies to the openAI format. This is useful so that our pipelines
always work on openAI format internally, both the current input pipeline
and the output pipeline.

The completion handler now really only does completion (previously it
was a really confusing class that did several things) and the adapter is
not hidden better in the litellmshim module.

To ship the PR faster, there are only two normalizers - OpenAI that just
passes the data through and Anthropic that uses the LiteLLM adapter.
Next, we'll add a llama.cpp normalizer to get rid of the `<im_start>`
tags and convert them into a properly formatted OpenAI message.