Optionally split the im_start/im_end tags that VLLM is using into individual messages
LiteLLM is able to pass through the messages as they are nominally in
the OpenAI format, but all the conversation was lumped into a single
message with the actual messges separated with the im_start/im_end tags.

Our pipeline needs to work on the individual messages, so let's add a
reusable normalizer that splits them. For now, we don't need to
denormalize as it looks like continue and litellm are OK with using
openAI everywhere later on.