Adds a streaming CodeCommentStep that adds a comment after every code block
We want to decorate code blocks streamed back to the user in a reply. To
do that, we first change the OutputPipelineInstance to buffer the full
reply. Then when the full reply is being streamed, we always check if
the chunk makes for a new code snippet. If yes, we print - for now, this
is to be changed - what kind of language the snippet was in.

We will use that as a base of checking if the code snippet contains any
malicious or archived packages.