Split out secret obfuscation into reusable classes
Instead of coding up the secret encryption directly in the Pipeline
step, let's split it out into a class of its own based on its own. The
actual method that changes the secret is pluggable, for encryption where
we need to get the secret value back we use the method we had used in
the pipeline step. For things like extracting packages from a code
snippet where we don't need to retrieve the original value we just
replace the secret with a fixed number of asterisks.