A temporary workaround to make Anthropic FIM working with Continue
We need to run the normalizer and denormalizer even if there is nothing
in the output pipeline for Anthropic because it uses a special format
for messages. However, we can't run the pipeline for all providers
because we are missing a normalizer/denormalizer for llama.cpp. We need
to add that, but let's add a temporary hack to get FIM with Anthropic
working.

Co-authored-by: Pankaj Telang <pankaj@stacklok.com>
Co-authored-by: Alejandro Ponce <aponcedeleonch@stacklok.com>