Add data types for pipeline routing, add route for OpenAI
We just had a simple if-else for pipeline routing. Instead, let's add a
data type that includes the paths and optionally a `target_url`. We can
use that to add a routing for OpenAI which we'll use for testing the
Copilot provider.

In the future we can make the routings pluggable.