fix: correct the way to get the last relevant user message for secrets
When showing new secrets detection, we were just picking from last
assistant message id. But in the case of aider, we need to pick more user
messages. So reuse the logic of the method to get the relevant user block,
to pick the index from the relevant user message and start counting
from there

Closes: #606