`make lint` now checks for format
This enables `make lint` to also run black but without persisting the
changes. This will hopefully allow developers to more easily detect if
their changes are not compliant with the linter settings.

This also removes the `make format` call from CI, which didn't actually
fail if linter issues were found.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>