Implement basic soft deletion for workspaces
this adds a `deleted_at` column to workspaces that implements a basic
soft-deletion mechanism. All relevant queries have been modified to reflect this.

At the moment, there is no hard deletion of workspaces; this will be
implemented in the future.

We also have no way of showing "archived" or "soft-deleted" workspaces.
This will come in due time.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>