Adds cache to cli-chat commands
Closes: #732

Copilot sends at least 2 requests with the same `last_user_message`.
Hence we execute the same command 2 times and reply to the last one.
The last behaviour will cause that if we create a Workspace with
Copilot the cli will respond that the workspace already exists.

This PR implements a workaround to cache the commands and it's outputs.
That way we can reply the same to subsequent requests sent by Copilot