Kick off provider endpoint CRUD structure and registration
This structure will handle all the database operations and turn that
into the right models.

Note that for provider endpoints we already have a way of setting these
via configuration, so this is taken into account to output some sample
objects that users can leverage.

Each provider will need to implement a `models` function which allows us
to auto-discover models for a provider.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>