Kick off provider endpoint CRUD structure
This structure will handle all the database operations and turn that
into the right models.

Note that for provider endpoints we already have a way of setting these
via configuration, so this is taken into account to output some sample
objects that users can leverage. We probably need to set some hardcoded
IDs sometime in the near future so we can reference these. But for now,
IDs for these are optional.

I also added a `source` field which allows us to tell where these
providers came from.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>