fix: remove checks for added into db because it was truncating content
We were inserting the first chunk that came from chat, and then skipping
to record the subsequent requests. But the content comes in chunks, so
we were not saving all the output from llm.
Remove this flag, and instead execute an UPSERT to insert/update content
if it was previously added