Move _get_base_url to the base provider
In order to properly support "muxing providers" like openrouter, we'll
have to tell litellm (or in future a native implementation), what server
do we want to proxy to. We were already doing that with Vllm, but since
are about to do the same for OpenRouter, let's move the `_get_base_url`
method to the base provider.