Add mux rules for match according to file and request type
Closes: #885 and #944

This PR introduces the necessary changes for:
1. Muxing to match on a specific filename (`main.py`) or file type (`.py`)
2. Muxing on a specific request type, i.e. chat or FIM

The matching is tightly coupled with the priority of the muxing rules.
A match will occur on the first possible rule.

Example:
Muxing rule 1 -> match all `.py` files -> go to `chat-gpt`
Muxing rule 2 -> match all `.md` files -> go to `ollama`

If in a request we recieve a `README.md` and a `main.py` file the
request will be routed to `chat-gpt` since it's the match with
the highest priority in the list.

To introduce above changes there were some minor changes:
1. Separate FIM detection into its own class. Before it used to be part of `BaseProvider`
2. Detect if a request is FIM or not before calling `process_request` method.

Both of them were necessary in order to re-use the logic to create the
matcher for request type