Workaround for litellm using a strict OpenAI provider for OpenRouter
The problem is that litellm uses the OpenAI provider for talking to
OpenRouter, but OpenRouter uses an OpenAI dialect - for example, the
OpenAI python API no longer allows you to POST payloads that contain
`post`, but those are often used with OpenRouter for FIM. The effect is
that FIM payloads from Continue are rejected by litellm.

To work around that, we add a FIM normalizer for OpenRouter that moves
prompt to messages, like we often do during normalization, but in this
normalizer we don't de-normalize but isntead pass on the payload with
`messages` to the completion.