feat: remove duplicated alerts
Sometimes the different client tools generate multiple requests when
the user requests a task. This generates what it looks as a duplicate
alert, but is not really a total duplicate as it belongs to different
request.
But for the user does not provide so much value to have it, so proceed
with deduplicating those alerts, based on the code snippet and details
of the alert

Also remove dogecoin regex as it's giving false positives

Closes: #875