Generate valid chunked response when returning short circuit answer
Prior to this change, this code would return a malformed HTTP response
with a chunk size, then HTTP header, then body, then the final chunk
size, final chunk then zero.

This changes the code to return the header before returning the code.
Furthermore, it does not count the size of the header in the first chunk
size.