Fix FIM on Continue. Have specific formatters for chat and FIM
Until now we had a general formatter on the way out of muxing.
This is wrong since sometimes the pipelines respond with different
format for chat or FIM. Such is the case for Ollama.

This PR separates the formatters and declares them explicitly so
that they're easier to adjust in the future.