Add integration tests with muxing
Replicate the current integration tests but instead of using
the specific provider URL, e.g. `/ollama` use the muxing
URL, i.e. `/v1/mux/`.

Muxing functionality should take care of routing the request
to the correct model and provider. For the moment we're
only going to test with the "catch_all" rule. Meaning, all the
requests will be directed to the same model.

In future iterations we can expand the integration tests to
check for multiple rules across different providers.