Make copilot integration tests pass on the replace-litellm branch (#1109)
* Fix copilot secret unredaction

The copilot provider always sends `cleanup_sensitive` set to `False` as it
manages the context itself. On streams where `finish_stream` was set to
`False` as well, we would have yielded the rest of the context buffer though
which would break secret unredaction.

To reproduce, ask Copilot to make a simple modification in a file
containing secrets so that it's forced to print the secrets back to you.

* Port recent copilot change to the new chunk API