Don't add the system prompt warning about secrets in the Secrets step (#1114)
This didn't matter when we were using the FIM normalizer as we would
have dropped the messages including the system message later, but since
we are passing the messages verbatim, let's not add the system message
when obfuscating secrets. Adding it means that the message is leaked to
the FIM tab-completion.