Pass `base_url` as argument through providers.
The `base_url` is set once at the provider level and is then retrieved
via instance variable, makes it harder to override in use cases like
muxing.

This change moves the `base_url` at the highest level possible,
i.e. the route handler, which is provider-specific and can then still
leverage the instance variable, while making it easier to override in
muxing handler by passing it explicitly to `process_request`.