Don't leak the PII system message into the FIM output
Applies to this branch only as it actually corectly handles the
messages. Similar to how we handle secrets, also PII adds a system
prompt that would be added to the messages and might affect the FIM
output where it might suggest a warning about leaked FIM.

I added some TODOs into the code because I don't love how the code is
structured but that change should be done in a separate work on main,
not in a topic branch.