Add option to run codegate in read-only mode
This will allow codegate to run without the v1 API and the pseudo-CLI.
while this is not super useful as it is today, in the future this will
allow us to run this as a sidecar container with AI agents.

The next step will be to add the ability to load workspace
configurations from files.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>