Workaround for system messages that contain a list of dicts
In the litellm `ChatCompletionMessage` that we use a system message is
defined as follows:
```
class OpenAIChatCompletionSystemMessage(TypedDict, total=False):
    role: Required[Literal["system"]]
    content: Required[Union[str, List]]
    name: str
```

So content can either be a string or a list. Our secret encryption code only
handled the string case. Since both cases will properly be handled by the
soon-to-be-coming rewrite, let's just add a workaround so that e.g. Cline with
Anthropic keeps working.

With the no-more-litellm branch, everything works as expected.

Fixes: #1207