Allow for dynamic workspace selection for muxing
This allows for the muxing endpoint to detect the `X-CodeGate-Workspace`
header in requests. This will allow CodeGate to search for that
workspace and use it's routing configuration explicitly.

If the a workspace is requested and it does not exist, a warning will be
issued and the active workspace will be defaulted to.

The use case for this is for background agents to be able to select a
relevant workspace and not need to disrupt the user's IDE session.

Note that this is not a persistent session and merely a dynamic runtime
setting.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>