Fix `codegate version` and similar commands.
While refactoring I removed three lines of coding managing
short-circuited requests. We short-circuit requests to implement
`codegate version`, `codegate workspace`, and similar commands.

Given we now have a provider-native representation of the messages, it
is necessary to produce the right message for the given request and
provider, so some code must be added in provider-specific modules to
handle that.

Fixes #1362